fix: 修复教学项目 P0 反模式（密钥/金额/JWT/CORS/上传/分页） (#47)
- 密钥：JwtSecret / SessionSecret / MoneySecret 走 config + 环境变量
  覆盖（JWT_SECRET / SESSION_SECRET / MONEY_SECRET），启动 fail-fast。
- JWT：EmailClaims 改用 PasswordDigest（bcrypt 哈希），不再带明文密码。
- 金额：Order.Money / DecryptMoney / PaymentDownReq.Money
  统一改 int64（分），strconv.FormatInt 替代 fmt.Sprintf("%f")。
- 支付鉴权：PaymentDownReq 仅保留 OrderId + Key；
  bossID/productID/num 全部从 order 取，新增已支付订单挡板与 Key 长度校验。
- 错误处理：补齐 product/payment 链路被忽略的 err。
- CORS：去掉 Allow-Origin: *，按 System.AllowedOrigins 白名单回显 Origin。
- Cookie：access/refresh cookie 增加 SameSite=Lax。
- 上传：替换 ioutil；按 MIME 校验 jpeg/png/gif/webp，限 5MB，
  权限 0644，文件名 sanitize。
- 分页：BasePage.Normalize() 限 PageNum/PageSize/LastId，
  新增 consts.MaxPageSize=100。
- 文档：新增 doc/MIGRATION.md，新增 config/locales/config.example.yaml。

+++ repository/db/model/user.go
@@ -64,13 +64,13 @@ func (u *User) EncryptMoney(key string) (money string, err error) {
 	return
 }
 
-// DecryptMoney 解密金额
-func (u *User) DecryptMoney(key string) (money float64, err error) {
+// DecryptMoney 解密金额，返回值单位为分。
+func (u *User) DecryptMoney(key string) (money int64, err error) {
 	aesObj, err := secret.NewAesEncrypt(conf.Config.EncryptSecret.MoneySecret, key, "", secret.AesEncrypt128, secret.AesModeTypeCBC)
 	if err != nil {
 		return
 	}
 
-	money = cast.ToFloat64(aesObj.SecretDecrypt(u.Money))
+	money = cast.ToInt64(aesObj.SecretDecrypt(u.Money))
 	return
 }
